@@ -1,25 +1,24 @@
 {
-  "name": "react-rails-api-project-template-client",
+  "name": "client",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:3000",
-  "engines": {
-    "node": "16.x"
-  },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.13.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "daisyui": "^2.31.0",
-    "react": "^17.0.2",
-    "react-bulma-components": "^4.1.0",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.4.1",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.2"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "autoprefixer": "^10.4.12",
+    "daisyui": "^2.33.0",
+    "postcss": "^8.4.18",
+    "react": "^18.2.0",
+    "react-daisyui": "^2.4.4",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "5.0.1",
+    "tailwindcss": "^3.2.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "PORT=4000 NODE_OPTIONS=--openssl-legacy-provider react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -41,10 +40,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.4.12",
-    "postcss": "^8.4.17",
-    "tailwindcss": "^3.1.8"
   }
 }
