@@ -1,13 +1,33 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+
 import './index.css';
-import App from './App';
-import './medium.ttf';
 
+// Routes
+/* Landing Pages */
+import Landing from './views/Landing';
+/* Auth Pages */
+// import Login from './views/Login';
+// import Register from './views/Register';
+/* Dashboard Pages */
+// import Dashboard from './views/Dashboard';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+    <BrowserRouter>
+      <Routes>
+        {/* Landing Pages */}
+        <Route path="/" element={<Landing />} />
+        
+        {/* Auth Pages */}
+        {/* <Route path="login" element={<Login /> } />
+        <Route path="register" element={<Register />} /> */}
+
+        {/* Dashboard Pages */}
+        {/* <Route path="dashboard" element={<Dashboard />} /> */}
+      </Routes>
+    </BrowserRouter>
+  </React.StrictMode>
 );
