@@ -1,10 +1,7 @@
-/** @type {import('tailwindcss').Config} */
 module.exports = {
-
-  
-  content: ["./src/**/*.{js,jsx,ts,tsx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [require("daisyui")],
+  content: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html', 'node_modules/daisyui/dist/**/*.js', 'node_modules/react-daisyui/dist/**/*.js'],
+  plugins: [require('daisyui')],
+  daisyui: {
+    themes: ["forest"]
+  }
 }
